test: remove snapshots on web as we don't want to re-test react-native-web lib (no specific changes)
